requirements.txt
requests==2.28.2
langchain==0.0.83
openai==0.26.5
python-telegram-bot==20.1
emoji==2.2.0
